fix: externalize layout in aith page
